+++ Config/conn.inc.php
@@ -1,5 +1,5 @@
 <?php
-define("DBHOST", "localhost");
-define("DBNAME", "phpcupcakes");
-define("DBUSER", "root");
-define("DBPASS", "");
+define("DBHOST", "193.203.166.181");
+define("DBNAME", "u265408878_phpcupcakes");
+define("DBUSER", "u265408878_phpcupcakes");
+define("DBPASS", "SUQ0;3i~W");