+++ Framework/Helpers/GetModels.php
@@ -22,7 +22,7 @@ namespace PhpCupcakes\Helpers;
                 $filePath = $folderPath . "/" . $file;
                 if (is_file($filePath) && substr($file, -4) === ".php") {
                     $className = substr($file, 0, -4);
-                    $fullClassName[] = $namespace. '\\' .$baseNamespace . '\\' . $className;
+                    $fullClassName = $namespace. '\\' .$baseNamespace . '\\' . $className;
                 }
             }
         }
@@ -33,20 +33,20 @@ namespace PhpCupcakes\Helpers;
     $myapps = FileFunctions::getFolders(ConfigVars::getDocRoot().'/Models');
         foreach ($myapps as $myapp) {
             if (file_exists(ConfigVars::getDocRoot().'/Models/'.$myapp.'/Models')) {
-        $models = self::getModelsNamespace(ConfigVars::getDocRoot().'/Models/'.$myapp.'/Models');
+        $models[] = self::getModelsNamespace(ConfigVars::getDocRoot().'/Models/'.$myapp.'/Models');
             }
         }
 
     $modules = FileFunctions::getFolders(ConfigVars::getDocRoot().'/Modules');
     foreach ($modules as $module) {
         if (file_exists(ConfigVars::getDocRoot().'/Modules/'.$module.'/Models')) {
-    $moremodels = self::getModelsNamespace(ConfigVars::getDocRoot().'/Modules/'.$module.'/Models');
+    $moremodels[] = self::getModelsNamespace(ConfigVars::getDocRoot().'/Modules/'.$module.'/Models');
         }
     }
     $plugins = FileFunctions::getFolders(ConfigVars::getDocRoot().'/Plugins');
     foreach ($plugins as $plugin) {
         if (file_exists(ConfigVars::getDocRoot().'/Plugins/'.$plugin.'/Models')) {
-    $evenmoremodels = self::getModelsNamespace(ConfigVars::getDocRoot().'/Plugins/'.$plugin.'/Models');
+    $evenmoremodels[] = self::getModelsNamespace(ConfigVars::getDocRoot().'/Plugins/'.$plugin.'/Models');
         }
     }
     if (isset($moremodels)) {